change int types to prevent possible errors in database/core

+++ MangosUtil/Parse.cs
@@ -122,7 +122,7 @@ namespace Mangos
 
         #endregion
 
-        public static ICondition Parse(ref int id, Expression<Func<ConditionFacade, bool>> expression)
+        public static ICondition Parse(ref ushort id, Expression<Func<ConditionFacade, bool>> expression)
         {
             var expr = _parser.Parse(expression.Body);
